requirements.txt
discord
pymongo
python-dotenv
DateTime
pillow
requests
pytz
